Update to Snapshot 3 of Pre-Indev (75 Minute Mark

+++ simplefileman.py
@@ -4,7 +4,7 @@ import logging
 
 """
 Tyler Lorence's Project #1 (Untitled)
-Version Pre-Indev (SNAPSHOT 2 - 60 MINUTE MARK)
+Version Pre-Indev (SNAPSHOT 3 - 75 MINUTE MARK)
 
 TODO: Add error handling with Try/Except.
 TODO: Add error logging with the logging module.
@@ -53,32 +53,43 @@ def create_file(filename):
     logger.info(f"Created the file {filename} successfully.")
 
 def change_directory(directory):
-    os.chdir(directory)
+    try:
+        os.chdir(directory)
+    except FileNotFoundError as e:
+        logger.error(f"Error attempting to change directory: Cannot find the directory {directory}")
 
 def make_directory(directory_name):
     os.mkdir(directory_name)
     logger.info(f"Created the directory {directory_name} successfully.")
 
 def remove_directory(directory_name):
-    os.rmdir(directory_name)
+    try:
+        os.rmdir(directory_name)
+    except FileNotFoundError:
+        logger.error(f"Error attempting to remove directory: Cannot find the directory {directory_name}")
 
 def read(file):
-    with open(file, "r") as f:
-        print(f"""\
-            ----- File {file} -----
-{f.read()}
-        """)
+    try:
+        with open(file, "r") as f:
+            print(f"""\
+                ----- File {file} -----
+    {f.read()}
+            """)
+    except FileNotFoundError:
+        logger.error(f"Error attempting to read file: Cannot find the file {file}")
 
 def remove_file(file):
     os.remove(file)
 
 def write_to_file(*contents):
-    with open(contents[0][0], "w+") as f:
-        logger.debug(f"\"*contents*\" parameter provided: {contents} / First index (at index 0) of contents parameter: {contents[0]}")
-        joined_contents = " ".join(contents[0][1:])
-        logger.debug(f"Joined contents via \"joined_contents\". type(joined_contents): {type(joined_contents)} / Joined Contents: {joined_contents}")
-        f.write(joined_contents)
-        # to_write = " ".join(contents[1:])
+    if os.path.isfile(contents[0][0]):
+        with open(contents[0][0], "w+") as f:
+            logger.debug(f"\"*contents*\" parameter provided: {contents} / First index (at index 0) of contents parameter: {contents[0]}")
+            joined_contents = " ".join(contents[0][1:])
+            logger.debug(f"Joined contents via \"joined_contents\". type(joined_contents): {type(joined_contents)} / Joined Contents: {joined_contents}")
+            f.write(joined_contents)
+    else:
+        logger.error(f"Error attempting to write to file: Cannot find the file {contents[0][1]}")
 
 
 while True:
@@ -114,9 +125,17 @@ while True:
                 else:
                     print("Invalid option! Please enter \"y\" or \"n\" for this prompt.")
     elif user_input[0].casefold() == "read" or user_input[0].casefold() == "more":
-        read(user_input[1])
+        try:
+            read(user_input[1])
+        except IndexError:
+            logger.error(f"Error attempting to read file: No file name was provided")
     elif user_input[0].casefold() == "write" or user_input[0].casefold() == "writeto":
-        write_to_file(user_input[1:])
+        try:
+            write_to_file(user_input[1:])
+        except IndexError:
+            logger.error(f"Error attempting to write to file: No filename was provided.")
+    elif user_input[0].casefold() == "help" or user_input[0].casefold() == "?":
+        print("This still needs to be added!")
     elif user_input[0].casefold() == "exit":
         exit(0)
     else:
